+++ src/NCalc.Core/Helpers/NumberType.cs
@@ -1,0 +1,10 @@
+namespace NCalc.Helpers;
+
+public enum NumberType
+{
+    Auto,
+    Double,
+    Decimal,
+    Int32,
+    Int64
+}