+++ src/NCalc.Parser/FloatingPointNumberType.cs
@@ -1,0 +1,7 @@
+namespace NCalc.Parser;
+
+public enum FloatingPointNumberType
+{
+    Double,
+    Decimal
+}