+++ src/NCalc.Parser/LogicalExpressionParserCacheKey.cs
@@ -2,14 +2,16 @@ namespace NCalc.Parser;
 
 internal readonly record struct LogicalExpressionParserCacheKey(
     bool AllowCharValues,
-    DefaultNumberType DefaultNumberType,
+    FloatingPointNumberType FloatingPointNumberType,
+    IntegerNumberType IntegerNumberType,
     ArgumentSeparator ArgumentSeparator,
     string CultureName)
 {
     public LogicalExpressionParserCacheKey(LogicalExpressionParserOptions options, CultureInfo culture)
         : this(
             options.AllowCharValues,
-            options.DefaultNumberType,
+            options.FloatingPointNumberType,
+            options.IntegerNumberType,
             options.ArgumentSeparator,
             culture.Name)
     {